Backend/requirements.txt
asgiref==3.8.1
Django==5.1.1
sqlparse==0.5.1
typing_extensions==4.12.2
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv